requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
appdirs==1.4.4
    # via
    #   hca-ingest
    #   requests-cache
attrs==21.4.0
    # via
    #   cattrs
    #   hca-ingest
    #   jsonschema
    #   requests-cache
cached-property==1.5.2
    # via kombu
cachetools==5.2.0
    # via
    #   -r requirements.in
    #   google-auth
cattrs==22.1.0
    # via
    #   hca-ingest
    #   requests-cache
certifi==2022.6.15
    # via
    #   hca-ingest
    #   requests
cffi==1.15.1
    # via
    #   cryptography
    #   hca-ingest
charset-normalizer==2.1.0
    # via
    #   hca-ingest
    #   requests
cryptography==37.0.4
    # via hca-ingest
dataclasses==0.6
    # via hca-ingest
et-xmlfile==1.1.0
    # via
    #   hca-ingest
    #   openpyxl
exceptiongroup==1.0.0rc8
    # via
    #   cattrs
    #   hca-ingest
google-api-core[grpc]==2.8.2
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-pubsub
    #   google-cloud-storage
google-api-python-client==2.56.0
    # via -r requirements.in
google-auth==2.10.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.2
    # via google-cloud-storage
google-cloud-pubsub==2.13.6
    # via -r requirements.in
google-cloud-storage==2.5.0
    # via -r requirements.in
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.3
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.56.4
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.4
    # via google-cloud-pubsub
grpcio==1.48.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.48.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
hca-ingest==2.3.0
    # via -r requirements.in
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via
    #   hca-ingest
    #   requests
importlib-metadata==4.12.0
    # via
    #   jsonschema
    #   kombu
importlib-resources==5.9.0
    # via jsonschema
jsonref==0.2
    # via hca-ingest
jsonschema==4.9.1
    # via -r requirements.in
kombu==5.2.4
    # via -r requirements.in
mergedeep==1.3.4
    # via hca-ingest
openpyxl==3.0.10
    # via hca-ingest
packaging==21.3
    # via -r requirements.in
pkgutil-resolve-name==1.3.10
    # via jsonschema
polling==0.3.2
    # via
    #   -r requirements.in
    #   hca-ingest
proto-plus==1.22.1
    # via google-cloud-pubsub
protobuf==4.21.5
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via
    #   cffi
    #   hca-ingest
pyjwt==2.4.0
    # via hca-ingest
pyparsing==3.0.9
    # via
    #   httplib2
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
pyyaml==6.0
    # via hca-ingest
requests[security]==2.28.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
    #   hca-ingest
    #   requests-cache
requests-cache==0.9.5
    # via hca-ingest
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
    #   hca-ingest
    #   url-normalize
sortedcontainers==2.4.0
    # via hca-ingest
typing-extensions==4.3.0
    # via
    #   cattrs
    #   importlib-metadata
    #   jsonschema
uritemplate==4.1.1
    # via google-api-python-client
url-normalize==1.4.3
    # via
    #   hca-ingest
    #   requests-cache
urllib3==1.26.11
    # via
    #   hca-ingest
    #   requests
    #   requests-cache
vine==5.0.0
    # via
    #   amqp
    #   kombu
xlsxwriter==3.0.3
    # via hca-ingest
zipp==3.8.1
    # via
    #   importlib-metadata
    #   importlib-resources
